I spearate the code into its own classes and that fixed the issue
